fix: package name (gs25 -> gs25)
