+++ pom.xml
@@ -10,7 +10,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
-        <exec.mainClass>gs25.hotel.reservation.management.system.HotelReservationManagementSystem</exec.mainClass>
+        <exec.mainClass>ds25.hotel.reservation.management.system.HotelReservationManagementSystem</exec.mainClass>
     </properties>
 
     <dependencies>
@@ -81,7 +81,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>gs25.hotel.reservation.management.system.HotelReservationManagementSystem</mainClass>
+                            <mainClass>ds25.hotel.reservation.management.system.HotelReservationManagementSystem</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
