+++ src/main/java/ds25/hotel/reservation/management/system/entity/user/UserGrade.java
@@ -1,0 +1,8 @@
+package ds25.hotel.reservation.management.system.entity.user;
+
+public enum UserGrade {
+    BRONZE,
+    SILVER,
+    GOLD,
+    PLATINUM
+}
