+++ src/main/java/ds25/hotel/reservation/management/system/entity/user/UserRole.java
@@ -1,0 +1,6 @@
+package ds25.hotel.reservation.management.system.entity.user;
+
+public enum UserRole {
+    ADMIN,
+    USER
+}
