+++ src/main/java/ds25/hotel/reservation/management/system/observer/Observable.java
@@ -1,0 +1,8 @@
+package ds25.hotel.reservation.management.system.observer;
+
+
+public interface Observable {
+    void registerObserver(Observer o);
+    void removeObserver(Observer o);
+    void notifyObservers();
+}
