+++ src/main/java/ds25/hotel/reservation/management/system/observer/Observer.java
@@ -1,0 +1,6 @@
+package ds25.hotel.reservation.management.system.observer;
+
+
+public interface Observer {
+    void update(Observable o, Object arg);
+}
