+++ src/main/java/ds25/hotel/reservation/management/system/service/hotel/HotelService.java
@@ -1,0 +1,4 @@
+package ds25.hotel.reservation.management.system.service.hotel;
+
+public class HotelService {
+}
