--- src/main/java/gs25/hotel/reservation/management/system/entity/hotel/HotelService.java
@@ -1,28 +1,0 @@
-package gs25.hotel.reservation.management.system.entity.hotel;
-
-public enum HotelService {
-    Bathtub,
-    bidet,
-    Shower,
-    bathrobe,
-    dryer,
-    WiFi,
-    television,
-    refrigerator,
-    AirConditioner,
-    ParkingAvailable,
-    SwimmingPool,
-    Breakfast,
-    terrace,
-    Sauna,
-    Restaurant,
-    Bar,
-    Fitness,
-    Buffet,
-    Amenities,
-    Desk24,
-    LendingNotebook,
-    Coffeeshop,
-    Business,
-    baggageStorage,
-}
