--- src/main/java/gs25/hotel/reservation/management/system/entity/user/UserGrade.java
@@ -1,8 +1,0 @@
-package gs25.hotel.reservation.management.system.entity.user;
-
-public enum UserGrade {
-    BRONZE,
-    SILVER,
-    GOLD,
-    PLATINUM
-}
