--- src/main/java/gs25/hotel/reservation/management/system/entity/user/UserRole.java
@@ -1,6 +1,0 @@
-package gs25.hotel.reservation.management.system.entity.user;
-
-public enum UserRole {
-    ADMIN,
-    USER
-}
