--- src/main/java/gs25/hotel/reservation/management/system/observer/Observable.java
@@ -1,8 +1,0 @@
-package gs25.hotel.reservation.management.system.observer;
-
-
-public interface Observable {
-    void registerObserver(Observer o);
-    void removeObserver(Observer o);
-    void notifyObservers();
-}
