--- src/main/java/gs25/hotel/reservation/management/system/observer/Observer.java
@@ -1,6 +1,0 @@
-package gs25.hotel.reservation.management.system.observer;
-
-
-public interface Observer {
-    void update(Observable o, Object arg);
-}
