--- src/main/java/gs25/hotel/reservation/management/system/service/hotel/HotelService.java
@@ -1,4 +1,0 @@
-package gs25.hotel.reservation.management.system.service.hotel;
-
-public class HotelService {
-}
